Moved Sky component to common components
