+++ src/components/common/Sky.jsx
